Arrow API Refactoring, extract implementation
